LLM_AI_trip/requirements.txt
diffusers==0.30.2
fastapi==0.115.0
huggingface_hub==0.24.6
langchain==0.3.0
langchain_cohere==0.3.0
langchain_core==0.3.5
pydantic==2.9.2
python-dotenv==1.0.1
torch==2.4.1+cu124
